fix code style  fixed

+++ src/AppBundle/Controller/OnlineController.php
@@ -33,22 +33,24 @@ class OnlineController extends BaseController
 
         $response = new Response('true');
         $response->headers->setCookie(new Cookie($cookieName, $uuid));
+
         return $response;
     }
 
     protected function generateGuid()
     {
-        if (function_exists('com_create_guid')){
+        if (function_exists('com_create_guid')) {
             return com_create_guid();
-        }else{
-            mt_srand((double)microtime()*10000);
+        } else {
+            mt_srand((float) microtime() * 10000);
             $charid = strtoupper(md5(uniqid(rand(), true)));
             $hyphen = chr(45);
             $uuid = substr($charid, 0, 8).$hyphen
                 .substr($charid, 8, 4).$hyphen
-                .substr($charid,12, 4).$hyphen
-                .substr($charid,16, 4).$hyphen
-                .substr($charid,20,12);
+                .substr($charid, 12, 4).$hyphen
+                .substr($charid, 16, 4).$hyphen
+                .substr($charid, 20, 12);
+
             return $uuid;
         }
     }
